+++ assets/cases/event/system-event/keyboard-event.scene.meta
@@ -1,5 +1,5 @@
 {
-  "ver": "1.1.27",
+  "ver": "1.1.30",
   "importer": "scene",
   "imported": true,
   "uuid": "0b782b0c-046f-496f-b84d-d5630adda7a4",
