+++ assets/cases/event/system-event/mouse-event.scene.meta
@@ -1,5 +1,5 @@
 {
-  "ver": "1.1.27",
+  "ver": "1.1.30",
   "importer": "scene",
   "imported": true,
   "uuid": "5d21cb2d-5036-411e-8e2e-b8f3b87d5ea5",
