+++ assets/cases/event/system-event/mouse-event.ts.meta
