+++ assets/shared-res/prefabs/NotSupportTip.prefab.meta
@@ -1,0 +1,13 @@
+{
+  "ver": "1.1.30",
+  "importer": "prefab",
+  "imported": true,
+  "uuid": "ef1f3d75-3557-486e-bab0-7171f2455a6b",
+  "files": [
+    ".json"
+  ],
+  "subMetas": {},
+  "userData": {
+    "syncNodeName": "NotSupportTip"
+  }
+}
